+++ rust-doomgeneric/engine/src/p_saveg.rs
@@ -102,7 +102,7 @@ pub enum SpecialThinkerClass {
     Endspecials,
 }
 pub const SAVEGAME_EOF: i32 = 0x1d;
-pub const VERSIONSIZE: i32 = 16;
+pub const VERSIONSIZE: usize = 16;
 pub fn temp_save_game_file(d_main: &DMainState, p_saveg: &mut PSavegState) -> String {
     if p_saveg.temp_savegame_filename.is_none() {
         p_saveg.temp_savegame_filename = Some(format!("{}temp.dsg", d_main.savegamedir));
@@ -368,10 +368,10 @@ fn saveg_read_player_t(state: &mut PSavegState, str: &mut Player) {
     str.health = saveg_read32(state);
     str.armorpoints = saveg_read32(state);
     str.armortype = saveg_read32(state);
-    for i in 0..(NUMPOWERS as usize) {
+    for i in 0..NUMPOWERS {
         str.powers[i] = saveg_read32(state);
     }
-    for i in 0..(NUMCARDS as usize) {
+    for i in 0..NUMCARDS {
         str.cards[i] = saveg_read32(state) != 0;
     }
     str.backpack = saveg_read32(state) != 0;
@@ -380,7 +380,7 @@ fn saveg_read_player_t(state: &mut PSavegState, str: &mut Player) {
     }
     str.readyweapon = weapontype_from_raw(saveg_read32(state));
     str.pendingweapon = weapontype_from_raw(saveg_read32(state));
-    for i in 0..(NUMWEAPONS as usize) {
+    for i in 0..NUMWEAPONS {
         str.weaponowned[i] = saveg_read32(state) != 0;
     }
     for i in 0..(NUMAMMO as usize) {
@@ -424,10 +424,10 @@ fn saveg_write_player_t(state: &mut PSavegState, str: &Player) {
     saveg_write32(state, str.health);
     saveg_write32(state, str.armorpoints);
     saveg_write32(state, str.armortype);
-    for i in 0..(NUMPOWERS as usize) {
+    for i in 0..NUMPOWERS {
         saveg_write32(state, str.powers[i]);
     }
-    for i in 0..(NUMCARDS as usize) {
+    for i in 0..NUMCARDS {
         saveg_write32(state, i32::from(str.cards[i]));
     }
     saveg_write32(state, i32::from(str.backpack));
@@ -436,7 +436,7 @@ fn saveg_write_player_t(state: &mut PSavegState, str: &Player) {
     }
     saveg_write32(state, str.readyweapon as i32);
     saveg_write32(state, str.pendingweapon as i32);
-    for i in 0..(NUMWEAPONS as usize) {
+    for i in 0..NUMWEAPONS {
         saveg_write32(state, i32::from(str.weaponowned[i]));
     }
     for i in 0..(NUMAMMO as usize) {
@@ -690,7 +690,7 @@ pub fn write_save_game_header(state: &mut GameState, description: &str) {
     let mut name_bytes = [0u8; 16];
     let copy_len = name.len().min(16);
     name_bytes[..copy_len].copy_from_slice(&name.as_bytes()[..copy_len]);
-    for &byte in name_bytes.iter().take(VERSIONSIZE as usize) {
+    for &byte in name_bytes.iter().take(VERSIONSIZE) {
         saveg_write8(&mut state.world.p_saveg, byte);
     }
     saveg_write8(&mut state.world.p_saveg, state.game.g_game.gameskill as u8);
@@ -727,7 +727,7 @@ pub fn read_save_game_header(state: &mut GameState) -> bool {
     for _ in 0..SAVESTRINGSIZE {
         saveg_read8(&mut state.world.p_saveg);
     }
-    for slot in read_vcheck.iter_mut().take(VERSIONSIZE as usize) {
+    for slot in read_vcheck.iter_mut().take(VERSIONSIZE) {
         *slot = saveg_read8(&mut state.world.p_saveg);
     }
     let version_name = format!("version {}", vanilla_version_code(&state.game.doomstat));
@@ -1020,7 +1020,7 @@ pub fn archive_specials(world: &mut World) {
                     .p_ceilng
                     .activeceilings
                     .iter()
-                    .take(MAXCEILINGS as usize)
+                    .take(MAXCEILINGS)
                     .any(|&entry| entry == Some(id));
                 if in_stasis {
                     let ceiling_id = world.p_tick.ceiling_payload(id);
